Amlogic ALSA configuration file fixes (#134)
* Add Khadas VIM3 configuration for HAOS 9.5

* Use new driver name for ALSA configuration

With Linux kernel commit bc387887ae22 ("ASoC: meson: implement
driver_name for snd_soc_card in meson-card-utils") [1] the Meson ASoC
driver passes the driver name (seemingly derrived from device tree
"compatible") to ALSA. This then reflects in /proc/asound/cards, which
now lists "axg-sound-card", instead of ODROID-N2 (derrived from the
device tree "model"):

Since Linux 5.16+
  # cat  /proc/asound/cards
   0 [ODROIDN2       ]: axg-sound-card - ODROID-N2
                        ODROID-N2

Linux 5.15 and earlier:
  # cat /proc/asound/cards
   0 [ODROIDN2       ]: ODROID-N2 - ODROID-N2
                        ODROID-N2

It seems that this string is used to find the ALSA configuration in
/usr/share/alsa/cards/.

Note that the same string is used in Khadas VIM3. But since the two
configuration files are the same, this actually deduplicate things.

[1] https://git.kernel.org/pub/scm/linux/kernel/git/torvalds/linux.git/commit/?id=bc387887ae22d6ed6439e83968b5be3443a8e57d

* Use new driver name for ALSA configuration for ODROID-C2

Similar to ODROID-N2, the driver name changed for ODROID-C2 as well.
Add a configuration file for the new driver name.